Preparing to remove the DataTable. Big Upgrade.

+++ VSFiles/DG/DataActions/SaveData.cs
@@ -39,9 +39,38 @@ namespace DG
             
         //If the folder does not exist, create it. if the folder does exist, do nothing.
         DirectoryInfo di = Directory.CreateDirectory(obtainDataDefinitions.TableDefinition.DataGeneratedPath);
-  
-        File.WriteAllText(filename, sb.ToString());
 
+        //Note that this does not close the file if it is open, but it does prevent an error message.
+         if (IsFileLocked(filename))
+         {
+            File.WriteAllText(filename, sb.ToString());
+         }
+
+      }
+
+      private  bool IsFileLocked(string filename)
+      {
+
+         FileInfo file = new FileInfo(filename);
+         
+         try
+         {
+            using (FileStream stream = file.Open(FileMode.Open, FileAccess.Read, FileShare.None))
+            {
+               stream.Close();
+            }
+         }
+         catch (IOException)
+         {
+            //the file is unavailable because it is:
+            //still being written to
+            //or being processed by another thread
+            //or does not exist (has already been processed)
+            return true;
+         }
+
+         //file is not locked
+         return false;
       }
    }
 }
